refactor: rewrite journeyGenerator.js as thin orchestration layer
